pyproject.toml
[project]
name = "docx-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "google-adk>=1.19.0",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.4",
    "lxml>=6.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
docx-analyze = "docx_analyzer.cli:main"
